feat: Modularize FastAPI implementation and clean up main file
